backend/requirements.txt
#needed libraries, too lazy to install them one by one lol

wait-for
fastapi
uvicorn
python-jose[cryptography]
pyjwt
python-dotenv
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary>=2.9.6
joblib
numpy
scikit-learn
pydantic[email]
mailersend
pydantic-settings>=2.0.0
alembic
pyotp
qrcode
pillow
pytest
requests
